\name{LLdecomp-internal}
\alias{f.data2impmat.R}
\alias{f.grenze2active.R}
\alias{f.makegraphtosplit2.R}
\alias{f.rang2finalmatrix.R}
\alias{f.split2.R}

\title{Internal LLdecomp functions}
\description{Internal LLdecomp  functions}
\usage{
f.data2impmat(data1,...)
f.grenze2active(mat01,grenze)
f.makegraphtosplit2(knoten,admatsym)
f.rang2finalmatrix(total)
f.split2(ugr,grenze=3)
}
\details{These are not to be called by the user.}
\keyword{internal}
